# packages/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Velocidoc PDF Generator API",
        "description": "Generate PDFs from Word templates with dynamic JSON data. Designed for Power Automate and n8n integrations.",
        "version": "1.0.0",
        "contact": {
            "name": "Velocidoc Support"
        }
    },
    "servers": [
        {
            "url": "https://your-api-host.com/v1",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:3000/v1",
            "description": "Local Development"
        }
    ],
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
    "paths": {
        "/generate": {
            "post": {
                "operationId": "generateDocument",
                "summary": "Generate PDF from Template",
                "description": "Renders a Word template with the provided JSON data and returns a PDF.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UniversalPayload"
                            },
                            "examples": {
                                "stored_template": {
                                    "summary": "Using a server-stored template",
                                    "value": {
                                        "template_id": "invoice.docx",
                                        "output_format": "pdf",
                                        "data": {
                                            "customer_name": "Acme Corp",
                                            "items": [
                                                { "description": "Widget A", "price": 10 },
                                                { "description": "Widget B", "price": 20 }
                                            ]
                                        }
                                    }
                                },
                                "byot": {
                                    "summary": "BYOT - Bring Your Own Template (from SharePoint)",
                                    "value": {
                                        "template": {
                                            "content": "UEsDBBQAAAAI... (base64-encoded DOCX)",
                                            "filename": "my-template.docx"
                                        },
                                        "output_format": "pdf",
                                        "data": {
                                            "customer_name": "Acme Corp"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Generated document",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "PDF conversion service unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/templates/validate": {
            "post": {
                "operationId": "validateTemplate",
                "summary": "Validate Template",
                "description": "Validates a DOCX template and extracts placeholder fields. Use this before generation to catch errors and discover required data fields.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidateTemplateRequest"
                            },
                            "example": {
                                "content": "UEsDBBQAAAAI... (base64-encoded DOCX)",
                                "data": {
                                    "customer_name": "Test"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateTemplateResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/templates": {
            "get": {
                "operationId": "listTemplates",
                "summary": "List Available Templates",
                "description": "Returns a list of available templates. Useful for dynamic dropdowns in Power Automate.",
                "responses": {
                    "200": {
                        "description": "List of templates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TemplateInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "API Key for authentication"
            }
        },
        "schemas": {
            "InlineTemplate": {
                "type": "object",
                "description": "Inline template for BYOT (Bring Your Own Template) - pass template directly from SharePoint/OneDrive",
                "required": ["content", "filename"],
                "properties": {
                    "content": {
                        "type": "string",
                        "format": "byte",
                        "description": "Base64-encoded DOCX file content"
                    },
                    "filename": {
                        "type": "string",
                        "description": "Original filename (e.g., 'invoice.docx')"
                    }
                }
            },
            "UniversalPayload": {
                "type": "object",
                "description": "Supports two modes: (1) template_id for server-stored templates, (2) template for BYOT. Provide one or the other, not both.",
                "required": [
                    "output_format",
                    "data"
                ],
                "properties": {
                    "template_id": {
                        "type": "string",
                        "description": "Server-stored template filename (e.g., 'invoice.docx'). Use this OR template, not both."
                    },
                    "template": {
                        "$ref": "#/components/schemas/InlineTemplate"
                    },
                    "output_format": {
                        "type": "string",
                        "enum": [
                            "pdf",
                            "docx"
                        ],
                        "description": "Desired output format"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "JSON data to merge into the template"
                    },
                    "options": {
                        "type": "object",
                        "properties": {
                            "header_text": {
                                "type": "string",
                                "description": "Optional header text"
                            },
                            "watermark": {
                                "type": "boolean",
                                "description": "Add watermark to output"
                            }
                        }
                    }
                }
            },
            "ValidateTemplateRequest": {
                "type": "object",
                "required": ["content"],
                "properties": {
                    "content": {
                        "type": "string",
                        "format": "byte",
                        "description": "Base64-encoded DOCX template content"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "Optional sample data to validate against"
                    }
                }
            },
            "ValidateTemplateResponse": {
                "type": "object",
                "properties": {
                    "valid": {
                        "type": "boolean",
                        "description": "Whether the template is valid"
                    },
                    "fields": {
                        "type": "array",
                        "items": { "type": "string" },
                        "description": "Extracted placeholder field names"
                    },
                    "errors": {
                        "type": "array",
                        "items": { "type": "string" },
                        "description": "Validation errors (if invalid)"
                    },
                    "warnings": {
                        "type": "array",
                        "items": { "type": "string" },
                        "description": "Non-fatal warnings"
                    }
                }
            },
            "TemplateInfo": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "required_fields": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    }
                }
            }
        }
    }
}